Version 0.0.5 (Alpha 5)
* Moved server folder from ~/.HomeKit-Bridge to the Indigo plugin
preferences folder on the recommendation of Jay from Indigo
* Fixed possible infinite loop issue where if starting the HB server
from the plugin it was not incrementing wait time, meaning you would
have to hard-stop the plugin to get it to recover if the HB server
didn't start up
* Changed service start and stop to quote the long path to the Indigo
prefs due to path complexity
* Added a failsafe on starting HB from the plugin that if it does not
start in the allotted time that it will issue an Unload command just to
make sure it isn't in memory and still attempting to start
* Added automatic HB server start on plugin startup as long as the
server option to auto start is checked - this allows for someone to
prevent an auto start by unchecking this box
* Added automatic HB server stop on plugin shutdown regardless of any
selection options because the API won't run when the plugin is stopped
anyway so why keep the HB server running?
* Saving a server config dialog will now update the address of the
server to reflect the HB port and username
* Added regular server process checking so that the running state of
the plugin device will reflect the running state of the service every
30-60 seconds (variable to keep CPU costs down)

Known Issues
---------------
* Integrated Homebridge server does not yet start, it reads the API and
identifies devices but does not actually start the full HB service.
* Action integration not fully coded
* Indigo server information is still being collected in the plugin
config but is no longer used or needed and should be removed since
don't talk to the Indigo API any longer
* The current API is not locked to Localhost but will need to be prior
to being publicly released for security purposes
* Editing an Action that was added to devices could result in the FILL
command being run when you save the device back to the server.  Current
work around is to select Action Group from the Object Type combobox
before editing Action Groups.
* Currently using a delay when we get a setCharacteristic to make sure
the value reports back in the JSON, but this should really be more
dynamic and should go into a loop until we get confirmation from the
device via Indigo events.  Since this is tricky it's slated for later
implementation because the workaround is fine for now.
* A failed Homebridge start can cause a minor race condition where HB
will continuously try to restart itself, the current solution to this
if it happens is to remove the serverId folder under ~/.HomeKit-Bridge
so that the restarts cannot be processed.  This is fine because the
plugin will regenerate that folder automatically when you turn on your
server device.  This has been countered by extra safeguards in server
startup and plugin shutdown but it's a HB issue that still needs
resolved.