Version 0.1.2 (Beta 1) - First Closed Beta Release
* Hello, I would like to introduce you to the new Fan device!  And it's
cool.  WOOT.
* Fixed issue with Insteon devices not reporting to HomeKit when they
have changed
* Fixed issue where server device would not automatically restart when
configuration was changed
* Added server delete routine so that if you delete a server it will
stop the server and remove the config folder
* Running multiple servers should now work fine but will need beta
tested to be certain, only limited alpha testing was performed

Known Issues
---------------
* Possible feedback loop because when I update HomeKit it causes you to
send a request to Indigo, which changes the device and that prompts the
plugin to then contact you to let you know that the device was updated.
 I’ll have to figure out a creative way to work around that.
* Possible minor race condition (it could be attributed to an
experimental version of HBB running) but sometimes commands from
HomeKit will cause a series of concurrent thread issues over multiple
plugins, indicating that Indigo is too busy to answer, this recovers
after a few seconds
* Indigo server information is still being collected in the plugin
config but is no longer used or needed and should be removed since
don't talk to the Indigo API any longer
* The current API is not locked to Localhost but will need to be prior
to being publicly released for security purposes
* A failed Homebridge start can cause a minor race condition where HB
will continuously try to restart itself, the current solution to this
if it happens is to remove the serverId folder under ~/.HomeKit-Bridge
so that the restarts cannot be processed.  This is fine because the
plugin will regenerate that folder automatically when you turn on your
server device.  This has been countered by extra safeguards in server
startup and plugin shutdown but it's a HB issue that still needs
resolved.
* Changing the port on a running server will result in the plugin
reporting that the port is in use when it's only in use by the
currently running server (resolve by stopping the server before
attempting to manually change any HB settings)
* If saving the server after adding an object the HomeKit name/type may
still show on the form but the Device/Action is set to Fill as it
should be