Version 0.5.0 (Beta 5)
Release Notes
==========
* Added new combobox action item in the server called "Delete and
Exclude" that allows you to not only delete an item from the list but
to also exclude it from all future lists (manageable from the menu),
this in direct response to [the complaint from Different
Computers](http://forums.indigodomo.com/viewtopic.php?p=154621#p154621)
regarding his frustrations with the way device inclusion works
* Added new combobox action item in the server called "Delete and Hide"
that allows you to not only delete an item from the list but to hide it
for the remainder of the time you have the server dialog open, allowing
you to re-fill as needed.  The cache will clear once the dialog is
cancelled or saved so that the object you hid will be available again
* Added new menu option to [Manage Hidden
Objects](https://github.com/Colorado4Wheeler/HomeKit-Bridge/wiki/Hidden-
Objects-Management)

Known Issues
---------------
* Able to add multiples of the same device but that will cause HomeKit
to crash if they are on the same server (perhaps even multiples since
it all dumps to HomeKit), need to enforce only adding one of any device
but also modify HomeKit to allow this under some circumstances like the
fan of an included thermostat since they need to be separate devices
* Need to remove API port from the plugin config and autodetect it
instead when building the server configuration
* Creating additional servers is not incrementing the user name and
then not showing up in Homebridge as a result (reported by Autolog
http://forums.indigodomo.com/viewtopic.php?p=154501#p154501)
* Complications spinning up a second server
(http://forums.indigodomo.com/viewtopic.php?p=154391#p154391).  BETA
TESTERS: PLEASE TEST ADDING MULTIPLE SERVERS.
* Brand new install with brand new server is still not auto starting
the server after config close (reported by Autolog, but he was on 0.1.1
and this was fixed in 0.1.2 so it may not be an issue after all, still
needs tested to absolutely certain so it was added to the testing issue
on Git)
* Possible feedback loop because when I update HomeKit it causes you to
send a request to Indigo, which changes the device and that prompts the
plugin to then contact you to let you know that the device was updated.
 I’ll have to figure out a creative way to work around that.
* Possible minor race condition (it could be attributed to an
experimental version of HBB running) but sometimes commands from
HomeKit will cause a series of concurrent thread issues over multiple
plugins, indicating that Indigo is too busy to answer, this recovers
after a few seconds
* The current API is not locked to Localhost but will need to be prior
to being publicly released for security purposes
* A failed Homebridge start can cause a minor race condition where HB
will continuously try to restart itself, the current solution to this
if it happens is to remove the serverId folder under ~/.HomeKit-Bridge
so that the restarts cannot be processed.  This is fine because the
plugin will regenerate that folder automatically when you turn on your
server device.  This has been countered by extra safeguards in server
startup and plugin shutdown but it's a HB issue that still needs
resolved.
* Changing the port on a running server will result in the plugin
reporting that the port is in use when it's only in use by the
currently running server (resolve by stopping the server before
attempting to manually change any HB settings)
* If saving the server after adding an object the HomeKit name/type may
still show on the form but the Device/Action is set to Fill as it
should be

Homebridge Issues
---------------
* Garage doors never even make a request to Indigo at all
* Change UUID creation to allow for multiple same devices, perhaps only
if a special key is passed

Wish List
---------------
* Autolog suggestion:
http://forums.indigodomo.com/viewtopic.php?p=154506#p154506
* Add a feature to read in HBB and Alexa items to build a cache of
already used alias names