Version 0.7.0 (Beta 7)
Release Notes
==========
* New device type added:
[Speaker](https://github.com/Colorado4Wheeler/HomeKit-Bridge/wiki/HomeKi
t-Model-Reference#speaker) - customized to work with Airfoil by default
if you choose an airfoil speaker from your list.  Unfortunately after
going through all the code to get this working it shows up in HomeKit
but is "not yet supported" so we'll have to wait for a HomeKit update -
but we'll be ready when HomeKit is.  In the meantime the Airfoil
control functionality has been added to Lightbulbs so you can use that
to Mute (on/off) and control volume (Brightness)
* New device type added:
[Microphone](https://github.com/Colorado4Wheeler/HomeKit-Bridge/wiki/Hom
eKit-Model-Reference#microphone) - I thought I would get clever and use
the microphone device since the definition is identical (mute, volume)
but, alas, this too is not yet supported
* New device type added: [Contact
Sensor](https://github.com/Colorado4Wheeler/HomeKit-Bridge/wiki/HomeKit-
Model-Reference#contactsensor) - this is the first device to actively
invert the onState of a device, since it's inverted in HomeKit
* Added Airfoil support despite the setbacks of using Speaker or
Microphone, Airfoil can be controlled via a Lightbulb device (On/Off is
Mute and Brightness is Volume).  Note that there were a few ways to
implement mute and I chose to go with connecting and disconnecting the
speaker rather than storing and retrieving the last speaker states
* Added failsafe for open port checking in case, for whatever reason, a
empty string is sent to the port checker as it caused an error [as
reported by
Coolcaper](http://forums.indigodomo.com/viewtopic.php?p=154850#p154850)
* Added failsafe for starting the server to make sure pin, port,
listenPort and username are all present and populated to prevent any
server startup errors on new server creation [as reported by
Coolcaper](http://forums.indigodomo.com/viewtopic.php?p=154854#p154854)
* Added invert option to the server configuration - you can read more
about it on the [Server Wiki
Page](https://github.com/Colorado4Wheeler/HomeKit-Bridge/wiki/Server-Dev
ice#including-your-indigo-objects) where it has been updated to explain
the new Invert option
* Added invert ability to Indigo devices that have an onState (dimmers,
relays, etc), this will only invert the onState if it's part of the
calculation
* Overhaul of HomeKit-to-Indigo commands, it will now wait for the
success of the command up to 25 seconds before aborting, please keep an
eye on the log for "Maximum time exceeded while setting the '' HomeKit
characteristic for..." errors in the log to see if this is going to
cause any problems, this overhaul was in direct result to
slow-to-respond devices such as Multi IO controllers that may need some
time before a door closes and engages the input sensor control
* Fixed typo in the device config where when you edit a device it would
not properly fill in the HomeKit device type that was saved with the
device
* Fixed bug with complication lookups that would cause an error when
selecting FILL or the divider line (oh, the things you don't think to
test that make you slap your head and say DOH!  This is why its in beta
testing)
* Fixed known issue: Garage doors using the Insteon Multi I/O kit don't
report back correctly because of the long delay in getting the Input1
to respond
* Fixed known issue: If saving the server after adding an object the
HomeKit name/type may still show on the form but the Device/Action is
set to Fill as it should be
* Known issue removed due to no reports of it: Possible feedback loop
because when I update HomeKit it causes you to send a request to
Indigo, which changes the device and that prompts the plugin to then
contact you to let you know that the device was updated.  I’ll have to
figure out a creative way to work around that.
* Known issue removed as it was another plugin causing this (and not
HKB or HBB): Possible minor race condition (it could be attributed to
an experimental version of HBB running) but sometimes commands from
HomeKit will cause a series of concurrent thread issues over multiple
plugins, indicating that Indigo is too busy to answer, this recovers
after a few seconds
* Filled wish list item: User requested devices: Door (Open/Close or
Contact) Sensor
* Filled wish list item: Dev device wishlist: Speaker

Known Issues
---------------
* FRESH INSTALLATION ISSUE: It seems fairly universal that the first
server you add does not work until you restart the plugin
* After all the code to get the Speaker (also Microphone) device
working it shows up in HomeKit but is "not yet supported" so we'll have
to wait for a HomeKit update - but we'll be ready when HomeKit is
* Need to remove API port from the plugin config and autodetect it
instead when building the server configuration
* Creating additional servers is not incrementing the user name and
then not showing up in Homebridge as a result (reported by Autolog
http://forums.indigodomo.com/viewtopic.php?p=154501#p154501)
* Complications spinning up a second server
(http://forums.indigodomo.com/viewtopic.php?p=154391#p154391).  BETA
TESTERS: PLEASE TEST ADDING MULTIPLE SERVERS.
* Brand new install with brand new server is still not auto starting
the server after config close (reported by Autolog, but he was on 0.1.1
and this was fixed in 0.1.2 so it may not be an issue after all, still
needs tested to absolutely certain so it was added to the testing issue
on Git)
* The current API is not locked to Localhost but will need to be prior
to being publicly released for security purposes
* A failed Homebridge start can cause a minor race condition where HB
will continuously try to restart itself, the current solution to this
if it happens is to remove the serverId folder under ~/.HomeKit-Bridge
so that the restarts cannot be processed.  This is fine because the
plugin will regenerate that folder automatically when you turn on your
server device.  This has been countered by extra safeguards in server
startup and plugin shutdown but it's a HB issue that still needs
resolved.
* Changing the port on a running server will result in the plugin
reporting that the port is in use when it's only in use by the
currently running server (resolve by stopping the server before
attempting to manually change any HB settings)

Homebridge Issues
---------------
* Need type and versByte to recognize

Wish List
---------------
* Dev device wishlist: Battery Service, Leak Sensor, Camera RTP Stream
Management, Lock Management
* Add ability to reverse On/Off on devices, many plugins do it opposite
of Indigo
* Autolog suggestion:
http://forums.indigodomo.com/viewtopic.php?p=154506#p154506
* Add a feature to read in HBB and Alexa items to build a cache of
already used alias names