Version 0.8.0 (Beta 8)
Release Notes
==========
* Changed behavior of Airfoil when using a Lightbulb HomeKit type so
that On/Off will connect/disconnect the Airfoil speaker instead of
using the Speaker control of Mute/Unmute which was exactly opposite
* Fixed bug on characteristic action execution where the thread debug
message indicating it was checking for a status change output to the
main console instead of thread debug
* Fixed bug on characteristic action execution where a run action would
error out if it had too many attempts, it should log the message rather
than error on the message [reported by
Jon](http://forums.indigodomo.com/viewtopic.php?p=154874#p154874)
* Fixed bug on devices that use the Invert On/Off option where the
display was correct but the commands would work only after timing out
since the commands were not also getting reversed properly
* Fixed bug on receiving characteristic changes from HomeKit where the
plugin may execute a characteristic change more than once under certain
circumstances, possibly reversing the action that was just executed
* Tested Lightbulb not working for Airfoil control [reported by
Jon](http://forums.indigodomo.com/viewtopic.php?p=154874#p154874) but
it was working (perhaps as result of the other bug fixes)

Known Issues
---------------
* FRESH INSTALLATION ISSUE: It seems fairly universal that the first
server you add does not work until you restart the plugin
* After all the code to get the Speaker (also Microphone) device
working it shows up in HomeKit but is "not yet supported" so we'll have
to wait for a HomeKit update - but we'll be ready when HomeKit is
* Need to remove API port from the plugin config and autodetect it
instead when building the server configuration
* Creating additional servers is not incrementing the user name and
then not showing up in Homebridge as a result (reported by Autolog
http://forums.indigodomo.com/viewtopic.php?p=154501#p154501)
* Complications spinning up a second server
(http://forums.indigodomo.com/viewtopic.php?p=154391#p154391).  BETA
TESTERS: PLEASE TEST ADDING MULTIPLE SERVERS.
* Brand new install with brand new server is still not auto starting
the server after config close (reported by Autolog, but he was on 0.1.1
and this was fixed in 0.1.2 so it may not be an issue after all, still
needs tested to absolutely certain so it was added to the testing issue
on Git)
* The current API is not locked to Localhost but will need to be prior
to being publicly released for security purposes
* A failed Homebridge start can cause a minor race condition where HB
will continuously try to restart itself, the current solution to this
if it happens is to remove the serverId folder under ~/.HomeKit-Bridge
so that the restarts cannot be processed.  This is fine because the
plugin will regenerate that folder automatically when you turn on your
server device.  This has been countered by extra safeguards in server
startup and plugin shutdown but it's a HB issue that still needs
resolved.
* Changing the port on a running server will result in the plugin
reporting that the port is in use when it's only in use by the
currently running server (resolve by stopping the server before
attempting to manually change any HB settings)

Homebridge Issues
---------------
* Need type and versByte to recognize

Wish List
---------------
* Dev device wishlist: Battery Service, Leak Sensor, Camera RTP Stream
Management, Lock Management
* Add ability to reverse On/Off on devices, many plugins do it opposite
of Indigo
* Autolog suggestion:
http://forums.indigodomo.com/viewtopic.php?p=154506#p154506
* Add a feature to read in HBB and Alexa items to build a cache of
already used alias names