Version 0.14.1 (Beta 14 Release 1)
Release Notes
==========
* NOTE: All known issues and wishlist items have been moved to [Git
issues](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues)
instead
* Added support for the Hue bulb plugin, if you use the hueBulb device
then the plugin will now support the color settings of the bulb.  This
is functional but still under construction so expect timeouts when you
change colors on the app until I dial that in, but it WILL change -
just don't go doing a thousand color changes until it's completed
* Renamed the device from being "HomeKit Bridge Server" to being
"HomeKit Accessory Server" to bring it more in line with what it is: a
HomeKit Accessory and also in pursuance of keeping the terms as user
friendly and universal as possible
* Changed max value of Color Temperature to 15,000
* Changed the description of "Fan Version 2" to simply "Fan" to avoid
confusion.  There is a Fan Version 1 but I don't see ever using it
since the V2 does everything the V1 did anyway
* Fixed fairly large bug where conversion from an Indigo type of
integer or float to a characteristic required value of the opposite
(float or integer) was using the wrong value for calculation, meaning
the result was always whatever the default for the characteristic was
(often zero) - if you had issues with a device not reporting correctly
please try it now to see if it resolved your issue
* Fixed minor bug on new server creation where if you do not add any
devices you might get an error when saving the server that "key device
not found in dict"
* Removed first revision service initializers (code cleanup)